+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+ - Restrict CPU profiling support to Node.js >=4.0.0
+
 ## 1.14.0
  - Add on-demand CPU profiling support.